Live inspectable inline charts with a near-fullscreen modal
render_plot now publishes a "chart" message block carrying the
validated spec and rows (capped at 5000 rows / 200KB; larger data falls
back to the PNG attachment). The web client renders the block live with
Observable Plot: a real SVG inline in the thread, dark-theme native,
with an HTML title and categorical legend chips, hover tooltips on the
primary mark by default, and an Expand button that opens the chart
re-rendered near-fullscreen (Esc, ✕, or backdrop click to close).

The pure chart engine moved from the adapters package to
@rakazo/core/plot so the server (PNG export) and browser (live render)
share one implementation; the adapters module re-exports it and keeps
only the sharp rasterizer. The PNG copy is still written to the bot's
workspace for export. History serializers describe chart blocks as
[chart: name] for models and mobile.

Verified live in the browser: an agent-requested multi-series line
chart rendered as an inline SVG, hover showed "product A · Month Mar ·
Revenue 48", Expand opened a 1126px live re-render, and Escape closed
it. 647 tests and 19/19 workspace checks pass.
