+++ packages/core/package.json
@@ -13,6 +13,11 @@
       "types": "./src/node/load-root-env.ts",
       "development": "./src/node/load-root-env.ts",
       "default": "./src/node/load-root-env.ts"
+    },
+    "./plot": {
+      "types": "./src/plot/plot-spec.ts",
+      "development": "./src/plot/plot-spec.ts",
+      "default": "./src/plot/plot-spec.ts"
     }
   },
   "scripts": {
@@ -21,11 +26,14 @@
   },
   "dependencies": {
     "@rakazo/contracts": "workspace:*",
-    "dotenv": "^17.2.1"
+    "dotenv": "^17.2.1",
+    "@observablehq/plot": "^0.6.17",
+    "d3-dsv": "^3.0.1"
   },
   "devDependencies": {
     "fast-check": "^4.3.0",
     "typescript": "^5.9.2",
-    "vitest": "^4.1.10"
+    "vitest": "^4.1.10",
+    "@types/d3-dsv": "^3.0.7"
   }
 }
